using duplex exclusively with looping sim
one channel = two channels one send one receive, send forwards to receive. A single channel would also behave the same way but this illustrates the general principle of where you put the processing loop or forwarding in a pipeline.

Sending to sender puts it in the transport's parent buffer, it is then processed somehow (eg copied) and then sent out another channel that someone is listening on. This is how the Conn works, creating a single form for multiple use cases. Really it's a read/writer/closer but just haven't coded it that way (yet!)