put node into session manager properly, change to sig API
switching over to schnorr next, just moving things around so the necessary hash and key parameters for a simple schnorr key validation without embedded recoverable signature. The ads must have the pubkey in them, for simplicity, they can then be parsed in and then the key is there to be used in a regular non-embedded signature.

DSA signatures overall are to be deprecated, and note to follow up on the equivalent type of Schnorr construction that enables recovery, just so it's available without having to use ecdsa to use this method.