removed redundant ad.Ad method from interface, replaced with Sign
The Splice method was literally the same as codec.Encode except no error return. Both wrong and redundant. Sign method was already primed to become interface so now ad.Ad is an interface that can do all the main things required of an Ad. Codec, plus sign and validate. This means that the implementation could now largely be purely abstract except for the conversion to and from the wire format.